optimized routes by shifting to the new router module
