+++ routers/index_router.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+# index end point for checking if it is working
+@router.get("/")
+def index():
+    return {"message": "Working."}
